Add bindep.txt file
bindep is a tool that allows expressing what the distro-package
requirements are for a source repository in human readable and machine
understandable format. With the file in place, one can see what's
missing for installation with:

  bindep -b

It also supports labelling depends, so in this case:

  bindep -b build

will show things needed both for install (unlabelled) and for build. A
quick way to make sure everything is set is:

  sudo apt-get install $(bindep -b build)

or

  sudo dnf install $(bindep -b build)

Assuming the machine in question has pip installed, the following will
install bindep globally:

  sudo pip install bindep

Or, alternately, if ~/.local/bin is in the user's PATH:

  pip install --user bindep

Some CI systems, such as zuul, have facilities to use bindep.txt files
to know what dependencies need to be installed before doing things.

It's worth noting that bindep does not attempt to provide any package
mapping, which is a choice that allows it to be a dead-simple package
listing tool.